requirements.txt
fastapi
opencv-python-headless
uvicorn
mediapipe
Pillow
numpy
python-multipart
